Initial checkin of FxGuiToolkit early access classes
All of the basic Eclipse and Maven support, and the five classes that are core to expanded support for tables in JavaFX, are part of this checkin. This code has been matched well with similar classes for Swing in the GuiToolkit library, and the two have converged to be as similar as possible. There are hundreds of other classes waiting to be ported to this library, but they will be grouped and prioritized based on ease of getting to consistent coding and naming standards with the existing toolkits and thoroughly fleshing out with standards-compliant javadocs, as the original version had little if any javadocs (though lots of regular code comments) or didn't follow Oracle's recommendations.