Rapid addition of preliminary versions of a lot of basic JavaFX customization that doesn't have external dependencies
Lots of standalone code for JavaFX enhancements, but rushed due to time-sensitive downstream client use. Not much for Javadocs yet, and some inconsistencies vs. the more established libraries at my GitHub page. This will be improved as time allows, and other stuff will be pulled in once it can be decoupled from external dependencies. Some of what is being checked in tonight had prior external dependencies that were refactored tonight so that the code could be checked in.