Not for use in production code yet; a rush job past midnight to get additions from late 2020 into the repository
It's been eons since I last did an update, and Eclipse didn't check everything in. At this super-late post-midnight hour, I can't finish the job until tomorrow, so this code should only be used as a reference vs. pulled into production projects. Mostly it is missing POM updates for new dependencies, including the just-published placeholder for an IO Toolkit (which might get renamed to Commons Toolkit). Includes extensions to ControlsFX Action Framework, which I may instead submit to that project for review (years in the making; it's way overdue for official review).