Major push to get most of the basic GUI control boilerplate code on-line
Almost everything is pushed now, with some stuff from the upcoming FxGraphicsToolkit as well (until I have time to split it off), but not yet my support for JavaFX Combo Box and Stage generalizations. Spinner tricks are included in this checkin, along with special ways of clamping and handling special Text Field behavior for both string and numeric data entry.