Massive checkin for charting, graphics, about box, splash screen, etc.
This is about as far as I can go without massive surgery on remaining code to checkin, and some Javadocs were dumbed down or turned into standard comments for now as the Javadocs requirements as enforced by Maven are far stricter than what was specified at the time much of this code was originally written (or borrowed from other open source projects).

Remaining code to check in, still has some unwanted dependencies on other libraries, but other stuff is fine if not for depending on those not-yet-published classes. So it's an incremental process, but going quickly.

Another thing to note is that some of this code depends on corrections and enhancements I made to ControlsFX, but that code isn't published yet (even though done years ago) as I need to figure out the correct process to use (e.g. should I fork ControlsFX or submit my changes and fixes for review by the development team, even though there's probably not much interest in devoting time to Java 8 based changes while the focus by that team is on later Java).

I have not had time yet to test the code that calls ControlsFX stuff to make sure it works OK without my changes in place. Also, I have not yet come up with the best way to expose resources such as CSS files, in a way that is friendly to Maven, Gradle, IntelliJ and Eclipse, so most of the CSS tags that show up in this code have no effect yet. That likely will be my next thrust with this library.

Some minor effort was exerted tonight to start organizing class methods in the way currently recommended by Oracle vs. the old-school way (from the C/C++ world) of alphabetizing them (this pre-dates modern IDE's that can dynamically present alphabetically ordered lists at will).