A few more related classes that were left out before (chart-related, login dialog, etc.)
Some of this needs further review as it elaborates upon lessons learned from semi-abandoned code in third-party libraries, and some is renamed and enhanced so as to not conflict with much more primitive versions in other libraries. Furthermore, an audit is needed so that any assume resources such as image icons, are gathered and added to this library.