Lots of new functionality, but one more pass is needed due to one Private API call
Unfortunately I ran out of time to investigate a ControlsFX Private API compiler error. I think this is in an area where I have long used my own custom extension of that lib, so I probably need to see how and where to make a form so I can properly post my revisions and maybe even find a way to contribute them (long overdue, but I haven't been able to keep up with the move past JavaFX 8 so far, as so much Private API was in use in SpreadsheetView).