Update GuiUtilities.java
Forgot to protect against resource not found when loading CSS. For now, we let a null pointer get captured by the existing exception catch for errors thrown by adding the CSS to the stylesheet list. In the future, we might design a special exception to throw and check for null directly.