Forgot to make toolkit's custom CSS optional for text fields, just combo boxes
As is often the case when a toolkit is extracted from a larger code base and made more generic, some use cases get overlooked or were hard-coded in the original context.

This checkin represents a thorough re-do of the toolkit's custom CSS application to custom controls, which was previously hard-coded to always be used for text fields but was optional for combo boxes.

Furthermore, the assumption was that one would skip if in a toolbar but otherwise would use the custom CSS, so the boolean logic for the previous flag was the opposite of what we need now, requiring massive code changes in downstream callers and derived classes plus factory methods.

Additionally, the base classes for custom text fields and combo boxes are now concrete vs abstract, allowing more flexibility in downstream clients and also making it possible to use XComboBox with enums.