Added custom controls for long values
An oversight earlier, as the controls for ints don't cover longs. Didn't do one for spinners though, as JavaFX Core doesn't have proper support for that unless we either use the Double based facilities and cast to Long or home-grow equivalents to the Integer based facilities used for the IntegerSpinnerStringConverter.