Corrected oversight on number based table cells
As with other control types, my original library revision naively provided a single Number based derivative class for table cells, forgetting that the template rules for generics don't allow this to be used across-the-board for floats and ints when the implementation was for doubles. Now there are separate implementations for the three main numeric types. I will probably add Long support as well.