Decoupled formatted text conversion of numbers from appending measurement unit string
This functionality was better off split into two methods, one calling the other, so that we can use the formatted text option in table cell contexts without appending the measurement unit string, which can annoy the user when they're in editing mode vs. viewing the presentation version of the value. They still want to be in localized number formatting though.