Fleshed out the menu system for stuff that typical document-based applications support
I forgot to post these upgrades earlier, after having added other stuff to the library that refers to the concept of a Project File as a Backing Store for application level data. This is generic enough that it should pertain to most such apps, and three menus are designed to be extended with custom features: Load, Import, Export.