Found a way to abstract an overridable enhanced JavaFX application class
This was a lot of work, to make an extended version of the JavaFX Application class that is still flexible enough to derive a custom application from it and only have to override a few methods if you want to diverge from the default choices. The lifespan and method calling order of Core JavaFX is adhered to, and a main() is provided for IDE testing, as well as a not-yet-complete skeletal DemoStage to launch. The JavaFX Primary Stage is only used for the Splash Screen and then dismissed, as it isn't very configurable even if you reset the scene content.