Fix tooltips due to Java 21 runtime exception on Java 8 libs
Running on Java 21 using pre-built Java 8 libraries, we get runtime exceptions for "can't set a value that is already bound" when setting the spinner tooltip on its editor vs. directly on the spinner.

At some point in the distant past when I first wrote this code, one had to set the tooltip on the editor vs. the spinner. As I am now only running this code in Java 21 runtime environments, I do not know if this edit breaks the runtime behavior in Java 8.