Update SceneGraphUtilities.java
Added new versions of the button and toggle button factory methods, that use the Action Framework and are icon-based vs. text-based, as the previous additions to this utility class were for buttons that do not need actions as they don't correspond to features exposed in menus or primary toolbars.

This is a rush job due to project needs, so the code will be cleaned up for common calling paths and better and more complete Javadocs soon, such as for explaining the purposes of each set of factory methods.