Add some icons for later, and improve some existing ones
This also adds two common Help Menu features that weren't in vogue yet at the time I first wrote this library. Knowledge Base is for a web-hosted set of articles that take the place of traditional all-in one user manuals accessed from the Help menu item. Report Issue is for providing a link to an issue tracking system in the browser for users to report problems.