Improved the colors for include/exclude options and added drag/drop table utils
Table Utilities now has a facility to add row drag/drop support to tables, using generics so that the code never has to be written again. Very careful edge casing is covered by this implementation so the user doesn't get stuck in an unfinished drag/drop sequence and has intuitive feedback the entire time they are selecting, dragging, and dropping a row (or not being able to due to invalid selection such as an empty row).