Forgot to provide default implementations for additional file handlers
Not all file action categories are handled by every app, so it's best to provide default no-op implementations for methods that would require access to a class variable (not possible in an interface, to avoid every implementing class needing to override methods that aren't relevant to their purpose. Downstream clients can always make their own abstract class to provide concrete overrides that are final and/or that access class variables.