Made a Print Manager to consolidate print handling
It is better to have a Print Manager as the implementation won't vary much if at all, and this allows the code to be used even if not deriving from XStage. It also pulls in the WebKit approach as an overall consolidated print manager.

Unfortunately, after updating Eclipse from a late 2020 build to the end of 2024 build, the previous checkin and this one both have compiler errors that weren't caught and will be handled later this weekend when I have time to bring the file handling overrides into sync with the new API that offloads those tasks to the FileActionHandler interface via default methods that can be overridden but which nevertheless need more parameters due to not having class variable references at hand.