Update TableUtilities.java
This is an attempt to avoid two edge cases for drag/drop in JavaFX tables, where one or more rows may get permanently deleted vs. moved, if dragging earlier in the table or well past the final table row. The real problem may be that this solution is incompatible with multi-select. There is another approach to drag/drop that is implemented at the level of TableView vs. TableRow, but we can't use it for POJO's as the JavaFX clipboard doesn't support basic objects vs. specific MIME types like images, text, etc.