Fixed some just-detected MRU issues
Previous usages of this library did not encounter end users who make liberal use of underscores in file names, so I had not noticed two issues that needed resolving today, one of which was simple and the other of which required deep dives into Core JavaFX and ControlsFX to see why underscores were stripped.

The simple fix was to check for whether a name already exists in the MRU list. Although this should never happen in deployed apps, during development stages there can be instability in preferences loading and saving and other factors that feed into the gathering of the initial MRU list, so it is best to check for duplicates after first stripping leading and trailing spaces from the name.

The trickier problem was the underscore getting stripped from filenames. As it turned out, only the first one got stripped, and it was due to multiple factors, one of which was the initial underscore being stripped before the placeholder MRU file numbers used as the menu item defaults (though also hidden and disabled until populated with an actual filename). This is because TextBinding.parseAndSplit() ignores mnemonic markers on text that has only one character -- a flaw in Core JavaFX.

Due to the initial underscore for the real mnemonic marker being stripped and not being replaced by my code that adds the filenames to the MRU items, the first underscore in the filename, if present, got detected as the mnemonic marker, was stripped from the filename as it appears in the menu item (but nowhere else, meaning it still worked for opening the file), and then underlining the next character (if the user enabled mnemonics using the OS shortcut key).