move remaining unused icons from other projects to core FX gui project so they're all together in one place and easy to find
I've decided it is best to have icons consolidated in one place, even if not used yet. These are all ones that I have used at one point or another, and it's easier to find them if they're all together in one place, but I may break off the icon resources into a separate project later.